Extract functionality of KillerThread and corresponding methods to PodReleaseService.
