+++ api/profiles/dev/application.properties
@@ -102,6 +102,7 @@ cluster.networks.config=
 cluster.cadvisor.port=4194
 cluster.cadvisor.request.period=15
 cluster.cadvisor.timeout=10
+cluster.unused.pod.release.rate=1000
 
 api.host=
 launch.script.url=
