+++ api/src/main/java/com/epam/pipeline/manager/preference/PreferenceValidators.java
@@ -156,10 +156,6 @@ public final class PreferenceValidators {
         return (pref, dependencies) -> StringUtils.isNumeric(pref) && Integer.parseInt(pref) >= x;
     }
 
-    public static BiPredicate<String, Map<String, Preference>> isLessThan(int x) {
-        return (pref, dependencies) -> NumberUtils.isNumber(pref) && Integer.parseInt(pref) < x;
-    }
-
     /**
      * A no-op validator, that is always true
      */
